+++ view/frontend/web/js/action/checkout-session-config-load.js
@@ -32,7 +32,8 @@ define([
     };
     return function (callback) {
         var cartId = customerData.get('cart')()['data_id'] || window.checkout.storeId;
-        // if (cartId !== getLocalStorage().get('cart_id')) {
+        var config = getLocalStorage().get('config') || false;
+        if (config && !config.checkout_payload.includes(document.URL)) {
             callbacks.push(callback);
             if (callbacks.length == 1) {
                 remoteStorage.get(url.build('amazon_pay/checkout/config')).done(function (config) {
@@ -43,8 +44,8 @@ define([
                     } while (callbacks.length);
                 });
             }
-        // } else {
-        //     callback(getLocalStorage().get('config'));
-        // }
+        } else {
+            callback(getLocalStorage().get('config'));
+        }
     };
 });